Move the shell exec helpers next to their only consumers
Squashed from:
- Move the shell exec helpers next to their only consumers
- Carry NodeAffinitySchedulingStrategy along with the moved exec helpers
- Stop patching command helpers on a module that no longer has them
