+++ examples/experimental/formal_math/single_round/kimina_wrapper.py
@@ -8,7 +8,8 @@ import requests
 from kimina_client import AsyncKiminaClient, CheckResponse
 from ray.util.scheduling_strategies import NodeAffinitySchedulingStrategy
 
-from miles.utils.misc import exec_command_cpu, get_free_port
+from miles.utils.external_utils.exec_command import exec_command_cpu
+from miles.utils.misc import get_free_port
 
 # TODO handle docker stop more gracefully later
 _KILL_PREVIOUS_KIMINA_DOCKER = bool(int(os.environ.get("MILES_KILL_PREVIOUS_KIMINA_DOCKER", "1")))
