+++ scripts/run_qwen3_4b_npu.py
@@ -1,6 +1,6 @@
 import os
 
-import miles.utils.misc as U
+import miles.utils.external_utils.exec_command as U
 from miles.utils.external_utils.command_utils import execute_train_npu
 
 MODEL_NAME = os.environ.get("MILES_SCRIPT_MODEL_NAME", "Qwen3-4B-Instruct-2507")
